+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "build": "webpack",
     "watch": "webpack --watch",
     "start": "webpack serve --open",
@@ -14,9 +14,14 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@jest/globals": "^29.6.2",
     "bootstrap-icons": "^1.10.5",
     "css-loader": "^6.8.1",
+    "fs": "^0.0.1-security",
+    "jsdom": "^22.1.0",
+    "jsdom-global": "^3.0.2",
     "lodash": "^4.17.21",
+    "path": "^0.12.7",
     "style-loader": "^3.3.3",
     "three": "^0.154.0",
     "webpack": "^5.88.2"
@@ -29,6 +34,7 @@
     "express": "^4.18.2",
     "hint": "^7.1.10",
     "html-webpack-plugin": "^5.5.3",
+    "jest": "^29.6.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
